[327. Count of Range Sum][Compile Error]committed by LeetCode Extension

+++ 327-Count-of-Range-Sum/solution.java
@@ -1,14 +1,14 @@
 public class Solution {
     public int countRangeSum(int[] nums, int lower, int upper) {
         int res = 0;
-        int prefixSum = 0;
-        TreeSet<Integer> set = new TreeSet<Integer>();
+        long prefixSum = 0;
+        TreeSet<Long> set = new TreeSet<Long>();
         set.add(0);
         for (int num : nums) {
             prefixSum += num;
             // cur - pre >= lower => pre <= cur - lower
             // cur - pre <= upper => pre >= cur - upper
-            Set<Integer> tmp = set.subSet(prefixSum - upper, true, prefixSum - lower, true);
+            Set<Long> tmp = set.subSet(prefixSum - upper, true, prefixSum - lower, true);
             res += tmp.size();
             set.add(prefixSum);
         }
